Switch to `uint64_t` instead of `long long` type
Use of strtoll() makes it impossible to enter values in the range
[LONG_LONG_MAX + 1; ULONG_LONG_MAX].

Also use of signed integers in operations that overflow invokes
undefined behaviour per C standard.

The interface even assumes 64 bits in the type, so I've used uint64_t
instead of `unsigned long long` which can have different width.

%lld and other printf-like formats accept `long long` or `unsigned long
long` and `uint64_t` is likely to be `unsigned long` or some other type
which is different from `unsigned long long` even if its width matches,
hence casts in several places.

+++ include/global.h
@@ -1,6 +1,8 @@
 #ifndef _GLOBAL_H
 #define _GLOBAL_H
 
+#include <stdint.h>
+
 #define MAX_IN 80
 #define INPUT_START 24
 
@@ -8,6 +10,6 @@
 
 void exit_pcalc(int);
 
-char *str_with_base_of_number(long long, int type);
+char *str_with_base_of_number(uint64_t, int type);
 
 #endif
